+++ ui/app/src/elements/components/applet-icon-badge.ts
