+++ ui/app/src/elements/components/join-group-card.ts
@@ -15,12 +15,12 @@ import {
   Dialog,
 } from "@scoped-elements/material-web";
 
-import { matrixContext } from "../context";
-import { MatrixStore } from "../matrix-store";
-import { sharedStyles } from "../sharedStyles";
+import { matrixContext } from "../../context";
+import { MatrixStore } from "../../matrix-store";
+import { sharedStyles } from "../../sharedStyles";
 import { query } from "lit/decorators.js";
 import { HoloHashMap, HoloIdenticon } from "@holochain-open-dev/utils";
-import { CreateWeGroupDialog } from "./create-we-group-dialog";
+import { CreateWeGroupDialog } from "../dialogs/create-we-group-dialog";
 import { SlTooltip } from "@scoped-elements/shoelace";
 import { ActionHash } from "@holochain/client";
