+++ ui/app/src/elements/components/notification-dot.ts
