+++ ui/app/src/elements/components/sidebar-button.ts
