+++ ui/app/src/elements/dialogs/create-applet-dialog.ts
@@ -9,18 +9,16 @@ import {
   Snackbar,
   Dialog,
   CircularProgress,
-  TextArea,
 } from "@scoped-elements/material-web";
 
-import { sharedStyles } from "../sharedStyles";
-import { AppletInfo } from "../types";
+import { sharedStyles } from "../../sharedStyles";
+import { AppletInfo } from "../../types";
 import { TaskSubscriber } from "lit-svelte-stores";
-import { MatrixStore } from "../matrix-store";
-import { matrixContext, weGroupContext } from "../context";
-import { DnaHash, EntryHash } from "@holochain/client";
-import { fakeEntryHash } from "@holochain-open-dev/utils";
+import { MatrixStore } from "../../matrix-store";
+import { matrixContext, weGroupContext } from "../../context";
+import { DnaHash } from "@holochain/client";
 
-export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
+export class CreateAppletDialog extends ScopedElementsMixin(LitElement) {
   @contextProvided({ context: matrixContext })
   _matrixStore!: MatrixStore;
 
@@ -33,21 +31,11 @@ export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
     () => [this._matrixStore, this.weGroupId]
   );
 
-  @property()
-  appletInstanceId!: EntryHash;
-
   @query("#applet-dialog")
   _appletDialog!: Dialog;
 
   @query("#installed-app-id")
   _installedAppIdField!: TextField;
-
-  @query("#subtitle-field")
-  _subtitleField!: TextArea;
-
-  @query("#description-field")
-  _descriptionField!: TextArea;
-
 
   @state()
   _dnaBundle: { hash: EntryHashB64; file: File } | undefined = undefined;
@@ -63,23 +51,22 @@ export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
   @state()
   _duplicateName: boolean = false;
 
+  @property()
+  _appletInfo: AppletInfo = {
+    title: "",
+    subtitle: "",
+    description: "",
+    devhubHappReleaseHash: new Uint8Array(0),
+    icon: undefined,
+  };
 
-  @state()
-  _fileBytes: Uint8Array | undefined = undefined;
-
-  open() {
+  open(appletInfo: AppletInfo) {
     this._appletDialog.show();
-  }
-
-  close() {
-    this._fileBytes = undefined;
-    this._subtitleField.value = "";
-    this._installedAppIdField.value = "";
-    this._descriptionField.value = "";
+    this._appletInfo = appletInfo;
   }
 
   get publishDisabled() {
-    return !this._fileBytes;
+    return !this._installedAppIdField || this._duplicateName;
   }
 
   checkValidity(_newValue, _nativeValidity) {
@@ -101,17 +88,14 @@ export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
     };
   }
 
-  async joinApplet() {
+  async createApplet() {
     (this.shadowRoot?.getElementById("installing-progress") as Snackbar).show();
     try {
-      console.log("CREATING APPLET FOR GROUP WITH ID: ", this.weGroupId);
-
-      await this._matrixStore.joinApplet(
+      const appletEntryHash = await this._matrixStore.createApplet(
         this.weGroupId,
-        this.appletInstanceId,
-        this._fileBytes, // compressed webhapp as Uint8Array
+        this._appletInfo,
+        this._installedAppIdField.value
       );
-      const appletEntryHash = this.appletInstanceId;
       (
         this.shadowRoot?.getElementById("installing-progress") as Snackbar
       ).close();
@@ -130,25 +114,6 @@ export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
       ).close();
       (this.shadowRoot?.getElementById("error-snackbar") as Snackbar).show();
       console.log("Installation error:", e);
-    }
-  }
-
-  // TODO! make typing right here
-  loadFileBytes(e: any) {
-    const files: FileList = e.target.files;
-
-    const reader = new FileReader();
-    reader.onload = (e) => {
-      console.log(e.target?.result);
-    }
-    console.log("FILES: ", files);
-    reader.readAsArrayBuffer(files[0]);
-    // TODO! make typing right here
-    reader.onloadend = (_e) => {
-      const buffer = reader.result as ArrayBuffer;
-      console.log("BUFFER: ", buffer);
-      const ui8 = new Uint8Array(buffer);
-      this._fileBytes = ui8;
     }
   }
 
@@ -184,19 +149,29 @@ export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
       ${this.renderInstallingProgress()}
 
       <mwc-dialog id="applet-dialog" heading="Add Custom Name">
-
-        <div>Upload the <b>same</b> .webhapp file as the person that installed it to the group:</div>
-
-        <input type="file" id="filepicker" accept=".webhapp" @change=${this.loadFileBytes}>
-        ${this._fileBytes
-            ? html``
-            : html`<div
+        <div class="column" style="padding: 16px; margin-bottom: 24px;">
+          <mwc-textfield
+            id="installed-app-id"
+            label="Custom Name"
+            required
+            outlined
+            autoValidate
+            value=${this._appletInfo.title}
+            @input=${() => this.requestUpdate()}
+            validateOnInitialRender
+            dialogInitialFocus
+            .validityTransform=${(newValue, nativeValidity) =>
+              this.checkValidity(newValue, nativeValidity)}
+          ></mwc-textfield>
+          ${this._duplicateName
+            ? html`<div
                 class="default-font"
                 style="color: #b10323; font-size: 12px; margin-left: 4px;"
               >
-                No file selected.
+                Name already exists.
               </div>`
-          }
+            : html``}
+        </div>
 
         <mwc-button
           slot="secondaryAction"
@@ -209,7 +184,7 @@ export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
           slot="primaryAction"
           dialogAction="close"
           label="INSTALL"
-          @click=${() => this.joinApplet()}
+          @click=${() => this.createApplet()}
         ></mwc-button>
       </mwc-dialog>
     `;
@@ -222,7 +197,6 @@ export class JoinFromFsDialog extends ScopedElementsMixin(LitElement) {
       "mwc-dialog": Dialog,
       "mwc-snackbar": Snackbar,
       "mwc-circular-progress": CircularProgress,
-      "mwc-textarea": TextArea,
     };
   }
